Change of URL to backend
Changed URL to directly access the database at dif03 (which is behind the firewall). If you want to run the frontend on localhost you have to be connected with vpn to NTNU.

Moved some end element characters up to the end of the previous line, mainly to avoid the syntax highlighting confusion.