Dynamic loading of related works
Added support for dynamic loading of related work when this tab is opened the first time. Works in the json data from the server now has a key for "related" if the work has related works. The value is initially set to null, and dynamically loaded data can be added as value after being loaded.